movie_booking_django/requirements.txt
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
pymongo==3.12.2
pytz==2024.1
